VPD-1472: Move to identifiers package
