+++ app/config/FrontendAppConfig.scala
@@ -17,7 +17,7 @@
 package config
 
 import com.google.inject.{Inject, Singleton}
-import models.{CredentialId, InternalId, VpdId}
+import models.identifiers.{CredentialId, InternalId, VpdId}
 import play.api.Configuration
 import play.api.i18n.Lang
 import play.api.mvc.RequestHeader
