+++ app/controllers/actions/ClaimEnrolmentAuthAction.scala
@@ -19,7 +19,7 @@ package controllers.actions
 import com.google.inject.Inject
 import config.FrontendAppConfig
 import controllers.routes
-import models.{InternalId, VpdId, GroupId}
+import models.identifiers.{GroupId, InternalId, VpdId}
 import models.requests.NoEnrolmentIdentifierRequest
 import play.api.Logging
 import play.api.mvc.*
