+++ app/models/requests/EnrolmentOptionalDataRequest.scala
@@ -16,8 +16,8 @@
 
 package models.requests
 
-import models.InternalId
 import models.enrolment.EnrolmentUserAnswers
+import models.identifiers.InternalId
 import play.api.mvc.{Request, WrappedRequest}
 
 case class EnrolmentOptionalDataRequest[A] (request: Request[A],
