+++ app/repositories/EnrolmentUserAnswersRepository.scala
@@ -17,8 +17,8 @@
 package repositories
 
 import config.FrontendAppConfig
-import models.InternalId
 import models.enrolment.EnrolmentUserAnswers
+import models.identifiers.InternalId
 import org.mongodb.scala.bson.conversions.Bson
 import org.mongodb.scala.model.*
 import play.api.libs.json.Format
